Fix compiler warnings in Lightning Network code
- Fix sign-compare warnings by casting to size_t in data.size() comparisons
- Fix GCC stringop-overread false positive by using std::lexicographical_compare
  instead of vector operator< for transaction output sorting

+++ src/lightning/bolt_transport.cpp
@@ -615,7 +615,7 @@ Result<LightningMessage> LightningMessage::Deserialize(const std::vector<uint8_t
     LightningMessage msg;
     msg.length = (static_cast<uint16_t>(data[0]) << 8) | static_cast<uint16_t>(data[1]);
 
-    if (data.size() < 2 + msg.length) {
+    if (data.size() < 2 + static_cast<size_t>(msg.length)) {
         return Result<LightningMessage>::Error("Incomplete message");
     }
